削除アイコンをサイドバーヘッダーからサイドコンテンツへ移動
