+++ src/components/shared/sidebar/hooks/useSideBarHeader.tsx
@@ -1,25 +1,10 @@
 import { useState } from 'react';
 
-type SideBarHeaderHooksType = {
-  handleDeleteClip: () => void;
-};
-
-export const useSideBarHeader = ({
-  handleDeleteClip,
-}: SideBarHeaderHooksType) => {
+export const useSideBarHeader = () => {
   const [isSearch, setIsSearch] = useState(false);
-  const [isModalOpen, setIsModalOpen] = useState(false);
-
-  const onClickDeleteOk = () => {
-    handleDeleteClip();
-    setIsModalOpen(false);
-  };
 
   return {
-    isModalOpen,
     isSearch,
-    onClickDeleteOk,
-    setIsModalOpen,
     setIsSearch,
   };
 };
